pyproject.toml
[project]
name = "watchdog-gov"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.76.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
